+++ lib/src/vr_player_created_callback.dart
@@ -1,0 +1,7 @@
+import 'package:vr_player/src/vr_player_controller.dart';
+import 'package:vr_player/src/vr_player_observer.dart';
+
+typedef VrPlayerCreatedCallback = void Function(
+  VrPlayerController controller,
+  VrPlayerObserver observer,
+);