+++ lib/src/vr_state.dart
@@ -1,0 +1,7 @@
+/// Enum for showing the player state
+enum VrState {
+  loading,
+  ready,
+  buffering,
+  idle,
+}