+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+Update package to Flutter 3.10.
+Dart 3 support
+Style fixes
+Bump Android Kaltura version to 4.28.3
+
 ## 0.1.5
 setVolume() method added. Player volume slider added.